fix: add nonce to uniquely identity units over CMR
In case of CMRs, the provider will see an obfuscated name for an unit,
which will cause the requiring to fail fetching its credentials from the
relation.

A requiring unit has to provide a unique string that will be used to
store the secret in the relation data, allowing this unit to fetch the
secret at a later time.

Rename <user provided suffix> to <requirer provided suffix>.