ci: let matrix tests run for a single interface from a PR branch (#292)
This PR moves the logic to invoke the matrix tests for an individual interface to a reusable workflow, and allows specifying the repository and branch to fetch the test and interface definitions from. This will allow the tests added by a PR to be (manually) run in the CI context before merging, instead of only being able to test them locally.