Add S3 v1 interface and update S3 v0 interface (#306)
This PR aims to introduce a newer version of `s3` interface (v1), which is going to be implemented by the `s3-integrator` on `2/` track going further. The new `s3-integrator` charm is added with features of automatic bucket creation and verification, as per the spec [DA171](https://docs.google.com/document/d/1eTPJ58MpDdToF9SicGr8XgXRdurio6XM3tvga0uXp4M/edit?usp=sharing).

The major changes in the interface is that now the sensitive information like `access-key` and `secret-key` are to be passed by encapsulating them in a Juju secret -- while other things being similar. The interface now also adds for support for both provider and requirer writing the version of the S3 lib they're using -- to enable us to have more flexibility in the future should we have the necessity to have conditional logic based on some specific version of charm lib.

This PR also aims to update some old and obsolete information on the `s3` v0 interface.